Actually process (most of) the log file instead
The current code stops log file analysis if $size-$pos is larger than the amount of lines (* 150 characters) to process in one run. This leads to only the first $lines being processed if a log file is larger than $lines * 150 bytes (which is usually the case). Changing the 'greater than' sign to a 'smaller than' sign fixes this issue. (However, the plugin will still nearly never analyze the most recent entries)